fix: satisfy AI providers quality gates
